refactor: 입출력 로직은 utils 폴더로 이동한다
