+++ README.md
@@ -48,7 +48,7 @@
 
 - [x] 우승자가 여러 명일 경우 쉼표(,)를 이용하여 구분한다.
 - [x] 자동차의 위치 값만큼 '-' 문자열로 반환한다.
-- [x]'{자동차 이름} : {위치만큼 '-' 표시}' 문구와 함께 전진할 때 자동차 이름과 위치를 함께 출력한다.
+- [x] '{자동차 이름} : {위치만큼 '-' 표시}' 문구와 함께 전진할 때 자동차 이름과 위치를 함께 출력한다.
 - [x] '최종 우승자: {자동차 이름}' 문구와 함께 우승한 자동차를 출력한다.
 - [x] 시도가 끝날 때마다 줄 바꿈 출력한다.
 - [x] '실행 결과' 문구를 출력한다.
@@ -69,7 +69,7 @@
   - [x] [예외]  자동차는 1대 이상 있어야 한다.
     - [x] 자동차 이름으로 빈 값이 들어올 수 없다.
 - [x] [입력] 시도할 횟수를 입력 받는다. - '시도할 횟수는 몇 회인가요?'
-  - [ ] [예외] 시도할 횟수는 양의 정수만 입력 가능하다.
+  - [x] [예외] 시도할 횟수는 양의 정수만 입력 가능하다.
 - [x] [출력] 실행 결과 문구를 출력한다. - '실행 결과'
 - [x] 각 자동차는 무작위 값이 4 이상이면 전진해서 위치가 1 증가하고 아니면 멈춰서 제자리에 있는다.
 - [x] 위에 문장을 시도할 횟수만큼 반복한다.
