+++ __tests__/OutputTest.js
@@ -1,4 +1,4 @@
-import { formatWinnerNames, generateProgressBar } from '../src/OutputUtils.js';
+import { formatWinnerNames, generateProgressBar } from '../src/utils/OutputUtils.js';
 
 describe('출력', () => {
   test('우승자가 여러 명일 경우 쉼표(,)를 이용하여 구분한다.', () => {
